fast-forward: Add option for no-ff merge commit generation
This change introduce a workflow setting (input) allowing to chose the
kind of generated merge commit.

This setting offers similar choices as those available under the GitHub
*Allow merge commit* option.

On pull-request events, the JSON keys set by the GitHub *Allow merge
commit* are available through the GitHub API, but not on PR comment
event. Hence, we have to set it in the workflow itself instead of
relying on the repository settings from GitHub.